Move SCIP and FSCIP tests to test_pulp file
